Adjust spacing and colors
