+++ app/src/components/Common/Button/Button.module.scss
@@ -19,6 +19,9 @@
   }
   &:focus {
     outline: 0;
+    background-color: var(--active);
+    color: var(--backgroundColor);
+    border-color: var(--backgroundColor);
   }
   &:disabled {
     background-color: var(--backgroundColor);
