+++ app/src/components/CreateEvent/CreateEventContainer.module.scss
@@ -8,4 +8,5 @@
   margin-top: 20px;
   display: flex;
   justify-content: space-between;
+  flex-direction: row-reverse;
 }
